overhaule/rationalise the testing system

+++ current_code/test/setupRun.sh
@@ -6,7 +6,7 @@ usage() {
   echo "  -m       don't do memory checking"
   echo "  -s       don't use any memory checking suppressions (exclusive to -m flag)"
   echo "  -c       don't do coverage checking"
-#  echo "  -b       don't rebuild app (may override -c)"
+  echo "  -b       don't rebuild app (may override -c)"
   echo "  <tests>  The tests you wish to run. Default to all tests."
   echo "             eg: 1*"
   echo "                 2_1 2_2 2_3"
@@ -16,10 +16,12 @@ usage() {
 #
 # Parse off all the parameters
 #
+
+NOBUILD=""
 RECORD=""
 SKIPMEMORY=""
 SKIPCOVERAGE=""
-while getopts ":hrmcs" flag 
+while getopts ":hrmcsb" flag 
 do 
   case $flag in 
     h)
@@ -41,6 +43,10 @@ do
     s)
       NOSUPPRESSIONS="-s";
     ;;
+    b)
+      NOBUILD="-b";
+      echo Will not \(re\)build the app.
+    ;;
   esac 
 done 
 if [ "$NOSUPPRESSIONS" != "" ]; then
@@ -59,7 +65,11 @@ if [ ! -f "/usr/bin/lcov" ]; then
   echo "Will skip analysing code coverage (package not available)."
 fi
 
-grep "^test" /etc/sane.d/dll.conf
+# So that everything else does not have to run as root (for testing), reset back later
+sudo chmod 757 /var/run
+sudo rm -f /var/log/opendias/opendias.log
+
+grep -q "^test" /etc/sane.d/dll.conf
 if [ "$?" -ne "0" ]; then
   echo enable sane testing by commenting in "test" in "/etc/sane.d/dll.conf"
   exit
@@ -68,40 +78,49 @@ fi
 #
 # Cleanup
 #
-rm -rf ../src/*.gcda ../src/*.gcno results/*
+rm -rf ../src/*.gcda ../src/*.gcno results
 mkdir -p results/coverage/
 mkdir -p results/resultsFiles/
 
 
-#
-# Compile, so that 'code coverage' analysis reports can be generated.
-#
-echo Cleaning ...
-cd ../
-make clean > test/results/buildLog.out
-cd test
-# if the file is here, then last time configure was run was in this script
-# so no need to re-do it.
-echo Configuring ...
-cd ../
-if [ "$SKIPCOVERAGE" == "-c" ]; then
-  ./configure CFLAGS=' -g ' &> test/results/buildLog2.out
-else
-  ./configure CFLAGS='--coverage' LIBS='-lgcov' &> test/results/buildLog2.out
-fi
-cd test
-# unfortunatly bash cannot support "&>>" - yet!
-cat results/buildLog2.out >> results/buildLog.out
-rm results/buildLog2.out
+if [ "$NOBUILD" == "" ]; then
+  #
+  # Compile, so that 'code coverage' analysis reports can be generated.
+  #
+  echo Cleaning ...
+  cd ../
+  make clean > test/results/buildLog.out
+  cd test
 
-# Will have to re-make, incase anything was changed in source.
-echo Making ...
-cd ../
-make &> test/results/buildLog2.out
-cd test
-# unfortunatly bash cannot support "&>>" - yet!
-cat results/buildLog2.out >> results/buildLog.out
-rm results/buildLog2.out
+  # if the file is here, then last time configure was run was in this script
+  # so no need to re-do it.
+  echo -n Configuring 
+  cd ../
+  if [ "$SKIPCOVERAGE" == "-c" ]; then
+    echo " (without coverage) ..."
+    ./configure CFLAGS=' -g ' &> test/results/buildLog2.out
+  else
+    echo " (with coverage) ..."
+    ./configure CFLAGS='--coverage' LIBS='-lgcov' &> test/results/buildLog2.out
+  fi
+  cd test
+  # unfortunatly bash cannot support "&>>" - yet!
+  cat results/buildLog2.out >> results/buildLog.out
+  rm results/buildLog2.out
+
+  # Will have to re-make, incase anything was changed in source.
+  echo Making ...
+  cd ../
+  make &> test/results/buildLog2.out
+  cd test
+  # unfortunatly bash cannot support "&>>" - yet!
+  cat results/buildLog2.out >> results/buildLog.out
+  rm results/buildLog2.out
+
+else
+  echo Skipping the build process
+fi
+
 
 #
 # Generate baseline (coverage) information
@@ -117,55 +136,49 @@ fi
 #
 echo Creating startup scripts ...
 if [ "$SKIPMEMORY" == "" ]; then
-  CODENAME=`lsb_release -c | cut -f2`
+  #CODENAME=`lsb_release -c | cut -f2`
   SUPPRESS=""
-#  for SUPP in `ls config/suppressions/$CODENAME/*`; do
-#    if [ -f $SUPP ]; then
-#      SUPPRESS="$SUPPRESS --suppressions=$SUPP"
-#    fi
-#  done
+  #if [ -d config/suppressions/$CODENAME ]; then
+  #  for SUPP in `ls config/suppressions/$CODENAME/*`; do
+  if [ -d suppressions ]; then
+    for SUPP in `ls suppressions/*`; do
+      if [ -f $SUPP ]; then
+        SUPPRESS="$SUPPRESS --suppressions=$SUPP"
+      fi
+    done
+  fi
   VALGRINDOPTS="--leak-check=full --leak-resolution=high --error-limit=no --tool=memcheck --num-callers=50 --log-file=results/resultsFiles/valgrind.out "
   GENSUPP="--gen-suppressions=all "
   VALGRIND="G_SLICE=always-malloc G_DEBUG=gc-friendly valgrind "
 #else
 #  VALGRIND="strace "
 fi
-echo touch runningTest > config/startAppCommands
-echo $VALGRIND $SUPPRESS $VALGRINDOPTS $GENSUPP ../src/opendias \&\> results/resultsFiles/appLog.out >> config/startAppCommands
-echo rm runningTest >> config/startAppCommands
-chmod 755 config/startAppCommands
+
+PWD=`pwd`
+echo $VALGRIND $SUPPRESS $VALGRINDOPTS $GENSUPP ../src/opendias -c $PWD/config/testapp.conf \> results/resultsFiles/appLog.out > config/startAppCommands
 
 
-#
-# Backup the data area
-#
-DTE=`date +%Y%m%d%H%M%S`
-if [ -d ~/.openDIAS ]; then
-  echo Backing up personal data ...
-  tar -czf conf/openDias_bkp.$DTE.tar.gz ~/.openDIAS
-  rm -rf ~/.openDIAS
-fi
 
-
+#######################################
 #######################################
 # Run automated tests
 echo Starting test harness ...
-./harness.sh $RECORD $SKIPMEMORY $@
+perl ./harness.pl $RECORD $SKIPMEMORY $@ 2> /dev/null
+#######################################
 #######################################
 
 
-
-#
-# Recover data area
-#
-if [ -f conf/openDias_bkp.$DTE.tar.gz ]; then
-  echo Restoring personal data ...
-  rm -rf ~/.openDIAS
-  OP=`pwd`
-  cd /
-  tar -xf conf/openDias_bkp.$DTE.tar.gz
-  cd $OP
-fi
+##
+## Recover data area
+##
+#if [ -f conf/openDias_bkp.$DTE.tar.gz ]; then
+#  echo Restoring personal data ...
+#  rm -rf ~/.openDIAS
+#  OP=`pwd`
+#  cd /
+#  tar -xf conf/openDias_bkp.$DTE.tar.gz
+#  cd $OP
+#fi
 
 
 #
@@ -173,12 +186,14 @@ fi
 #
 if [ "$SKIPCOVERAGE" == "" ]; then
   echo Creating run coverage information ...
+  echo Creating run coverage information ... >> results/buildLog.out
   lcov -c -d ../src -o results/coverage/app_test.info >> results/buildLog.out
   lcov -a results/coverage/app_base.info -a results/coverage/app_test.info -o results/coverage/app_total.info >> results/buildLog.out
   genhtml -o results/coverage results/coverage/app_total.info >> results/buildLog.out
 fi
 
-rm config/startAppCommands
+#rm config/startAppCommands
 
+sudo chmod 755 /var/run
